Support custom mounts and environment variables (#194)
Part of: https://github.com/creek-service/creek-system-test/issues/172

Enhance the executor to support user supplied mounts and environment variables to add to Docker containers.  This will be required to support line coverage.

The same mechanism can also be used to handle service debugging.  This moves the construction of the `JAVA_TOOL_OPTIONS` env var out of the executor and into the Gradle plugin.
This has benefits:
 1. The executor becomes agnostic to debugging.
 2. Different build plugins could support different ways to debug a service.
 2. The build plugin can create a directory containing the required agent and mount on container, removing the requirement that the agent be included in the service's image.